requirements.txt
absl-py==0.9.0
altgraph==0.17
appdirs==1.4.4
astunparse==1.6.3
attrs==19.3.0
audioread==2.1.8
bokeh==1.4.0
cachetools==4.1.1
cairocffi==1.2.0
CairoSVG==2.5.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
clldutils==3.5.2
colorama==0.4.3
colorlog==4.2.1
cryptography==3.3.1
cssselect2==0.4.1
csvw==1.7.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
face-alignment==1.0.1
Flask==1.1.2
Flask-Cors==3.0.10
freetype-py==2.2.0
future==0.18.2
gast==0.3.3
glfw==2.0.0
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgui==1.3.0
inflect==4.1.0
intervaltree==3.1.0
irc==19.0.1
irc-bot==1.0.35
isodate==0.6.0
itsdangerous==1.1.0
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.1
jaraco.logging==3.0.0
jaraco.stream==3.0.0
jaraco.text==3.2.0
Jinja2==2.11.2
joblib==0.16.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
librosa==0.8.0
llvmlite==0.34.0
lxml==4.6.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.0
more-itertools==8.5.0
networkx==2.4
numba==0.51.0
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.3.0.36
opt-einsum==3.3.0
packaging==20.4
pandas==1.0.5
pefile==2019.4.18
phonemizer==2.2.1
Pillow==7.2.0
pooch==1.1.1
praw==7.1.0
prawcore==1.5.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.10
pylzma==0.5.0
PyOpenGL==3.1.5
pyparsing==2.4.7
pypng==0.0.20
pyswf==1.5.4
PyTexturePacker==1.1.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-resize-image==1.1.19
pytz==2020.1
PyWavelets==1.1.1
pywin32-ctypes==0.2.0
PyYAML==5.3.1
regex==2020.7.14
reportlab==3.5.56
requests==2.24.0
requests-oauthlib==1.3.0
resampy==0.2.2
rfc3986==1.4.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.4.1
segments==2.1.3
six==1.15.0
sklearn==0.0
sortedcontainers==2.3.0
SoundFile==0.10.3.post1
svglib==1.0.1
tabulate==0.8.7
tag==0.5
tempora==4.0.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.7.17
tinycss2==1.1.0
torch==1.5.1
torchvision==0.6.1
tornado==6.0.4
tqdm==4.48.0
-e git+https://github.com/mozilla/TTS.git@6d6dca00df024412b8421530de60de2dd2d64ad9#egg=TTS
Unidecode==1.1.1
update-checker==0.18.0
uritemplate==3.0.1
urllib3==1.25.10
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
youtube-dl==2020.12.7
